requirements.txt
suds-jurko
six